LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-009-expected.txt


PASS Should use the last fallback position initially
FAIL Should use the third fallback position with enough space right assert_equals: Anchored element should be at the right of anchor expected 650 but got 700
FAIL Should use the second fallback position with enough space below assert_equals: Anchored element should be at the left of anchor expected 650 but got 800
FAIL Should use the first fallback position with enough space right and below assert_equals: Anchored element should be at the right of anchor expected 199 but got 700
